+++ mrv_vectors.h
@@ -6,7 +6,6 @@
 #define MRV_VECTORS_H
 
 #include "mrm_misc.h"
-#include <stdlib.h>
 
 /*
  * vectors!
@@ -30,18 +29,18 @@
 
 typedef struct {
 	unsigned char *arr;
-	size_t stride;
-	size_t len;
-	size_t capacity;
+	uint stride;
+	uint len;
+	uint capacity;
 } MrvVector;
 
-MrvVector *mrv_create(size_t capacity, size_t stride);
+MrvVector *mrv_create(uint capacity, uint stride);
 
 void mrv_destroy(MrvVector *vec);
 
 void mrv_free(MrvVector *vec);
 
-void mrv_init(MrvVector *vec, size_t capacity, size_t stride);
+void mrv_init(MrvVector *vec, uint capacity, uint stride);
 
 Err mrv_clear(MrvVector *vec);
 
@@ -52,7 +51,7 @@ Err mrv_realloc_to_fit(MrvVector *vec);
 
 Err mrv_pop(MrvVector *vec);
 
-void *mrv_get_idx(MrvVector *vec, size_t n);
+void *mrv_get_idx(MrvVector *vec, uint n);
 
 void *mrv_get_item(MrvVector *vec, void *item);
